requirements.txt
python-dotenv
pycodestyle
autopep8
py-cord
beautifulsoup4
CurrencyConverter
matplotlib
pandas
bar_chart_race
pymongo
saucenao-api